requirements.txt
transformers
torch
huggingface_hub
mastodon.py
python-dotenv
requests